More money for the people!

+++ src/Interface/Text.h
@@ -22,6 +22,7 @@
 #include "../Engine/Surface.h"
 #include <vector>
 #include <string>
+#include <stdint.h>
 
 namespace OpenXcom
 {
@@ -60,9 +61,9 @@ public:
 	/// Cleans up the text.
 	~Text();
 	/// Formats an integer value as number with separators.
-	static std::wstring formatNumber(int value, std::wstring currency = L"");
+	static std::wstring formatNumber(int64_t value, std::wstring currency = L"");
 	/// Formats an integer value as currency.
-	static std::wstring formatFunding(int funds);
+	static std::wstring formatFunding(int64_t funds);
 	/// Formats an integer value as percentage.
 	static std::wstring formatPercentage(int value);
 	/// Sets the text size to big.
